Remove math functions from component
